WokkahLearn/requirements.txt
# Core Django and Database
Django==4.2.7
psycopg2-binary==2.9.7
django-environ==0.11.2

# REST API
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0

# Real-time features
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1

# AI Integration
openai==1.3.6
anthropic==0.8.1
langchain==0.0.340
langchain-openai==0.0.2

# File handling and storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.29.7

# Celery for async tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Payment integration
stripe==7.8.0

# WebRTC and real-time collaboration
django-channels-presence==1.0.0

# Content and media processing
python-magic==0.4.27
moviepy==1.0.3

# Analytics and monitoring
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Security and authentication
django-allauth==0.57.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Testing
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Development tools
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0

# Documentation
drf-spectacular==0.26.5

# Background tasks and scheduling
APScheduler==3.10.4

# WebSocket support for real-time features
websockets==12.0

# Code execution sandbox
docker==6.1.3

# Machine Learning for personalization
scikit-learn==1.3.2
numpy==1.25.2
pandas==1.5.3

# Content recommendation engine
surprise==1.1.3

# Video processing for course content
ffmpeg-python==0.2.0

djangorestframework-simplejwt

django-channels

django-allauth

drf-spectacular
django-celery-beat
django-celery-results